refactor to `frame_averaging.py`
